requirements.txt
Django==1.8.3
Pillow==2.9.0
argparse==1.2.1
django-appconf==1.0.1
django-ckeditor==4.5.1
django-geoposition==0.2.2
django-taggit==0.16.3
psycopg2==2.6.1
six==1.9.0
wsgiref==0.1.2
